change PositiveInegerField to BigItegerField for max-capacity field in Profile

+++ members/models.py
@@ -168,7 +168,7 @@ class Profile(models.Model):
     member = models.OneToOneField(Member, on_delete=models.CASCADE)
     slug = models.SlugField()
     last_marshmallow_allocation = models.DateTimeField(default=timezone.now)
-    media_capacity = models.PositiveIntegerField(default=DEFAULT_MEMBER_STORAGE)
+    media_capacity = models.BigIntegerField(default=DEFAULT_MEMBER_STORAGE)
     text = models.TextField(
         blank=True,
         null=True,
